implement single responsibility using class Method From API
