+++ PracticalTest.BusinessObjects/SportEventsValidator.cs
@@ -1,9 +1,4 @@
 ﻿using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 namespace PracticalTest.BusinessObjects
 {
     public class SportEventsValidator : AbstractValidator<SportEvents>
