+++ PracticalTest.DAL/Implement/SportEventsRepository.cs
@@ -1,22 +1,15 @@
-﻿using Microsoft.Net.Http.Headers;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using PracticalTest.BusinessObjects;
 using PracticalTest.Contracts;
-using PracticalTest.DTO;
 using PracticalTest.DTO.Create;
-using System.IdentityModel.Tokens.Jwt;
-using System.Net.Http.Headers;
-using System.Net.Http.Json;
-using System.Text;
-using System.Text.Json.Nodes;
 
 namespace PracticalTest.DAL
 {
     public class SportEventRepository : ISportEventsRepository
     {
-        private readonly ILoginToAPI _loginToAPI;
+        private readonly IMethodFromAPI _loginToAPI;
 
-        public SportEventRepository(ILoginToAPI loginToAPI)
+        public SportEventRepository(IMethodFromAPI loginToAPI)
         {
             _loginToAPI = loginToAPI;
         }
@@ -42,7 +35,6 @@ namespace PracticalTest.DAL
             var insert = _loginToAPI.InsertAsync(sportEventsCreateAPIDTO);
             var get = insert.ToString();
             var result = JsonConvert.DeserializeObject<SportEvents>(get);
-            return;
         }
 
         public void Edit(int id, SportEventsCreateAPIDTO sportEventsCreateAPIDTO)
@@ -52,7 +44,7 @@ namespace PracticalTest.DAL
 
         public async void Delete(int id)
         {
-            await _loginToAPI.DeleteAsync(id);
+            await _loginToAPI.DeleteSportEventAsync(id);
         }
 
     }
