+++ PracticalTest.DTO/SportEventsDTO.cs
@@ -1,9 +1,4 @@
 ﻿using PracticalTest.BusinessObjects;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PracticalTest.DTO
 {
